Upd. Integrations. Refactoring BuddyPress handler. (#504)
* Upd. Integrations. Refactoring BuddyPress handler.

* update exception rule
