+++ inc/cleantalk-admin.php
@@ -28,16 +28,6 @@ add_action('comment_approved_to_unapproved', 'apbct_comment__remove_meta_approve
 add_action('comment_spam_to_unapproved', 'apbct_comment__remove_meta_approved', 10, 1);
 add_action('comment_trash_to_unapproved', 'apbct_comment__remove_meta_approved', 10, 1);
 
-// Handle buddyPress users manage hooks to cathch spam/not spam feedback
-if ( apbct_is_plugin_active('buddypress/bp-loader.php') ) {
-    add_action('make_spam_user', 'apbct_send_buddypress_user_feedback');
-    add_action('make_ham_user', 'apbct_send_buddypress_user_feedback');
-    // Show admin notice on the users page
-    if (!empty($apbct->data['bp_feedback_message'])) {
-        add_action('admin_notices', 'apbct_buddypress_user_feedback_show_admin_notice', 998);
-    }
-}
-
 /**
  * Crunch for Anti-Bot
  * Hooked by 'admin_head'
@@ -1376,68 +1366,6 @@ function apbct_comment__send_feedback(
     if ( ! $direct_call ) {
         ! empty($result) ? die($result) : die(0);
     }
-}
-
-/**
- * WP hook action. Adds an admin notice if feedback from buddypress hooks collected and prepared to send.
- * @return void
- */
-function apbct_buddypress_user_feedback_show_admin_notice()
-{
-    global $apbct;
-    // second check if message persists
-    if (!empty($apbct->data['bp_feedback_message'])) {
-        $html = '<div class="notice notice-success is-dismissible">
-                                    <p>' . $apbct->data['bp_feedback_message'] . '</p>
-                                </div>';
-        echo Escape::escKsesPreset(
-            $html,
-            'apbct_response_custom_message'
-        );
-        // clear message to prevent next show
-        $apbct->data['bp_feedback_message'] = null;
-        $apbct->saveData();
-    }
-}
-
-/**
- * WP hook action. Handles BuddyPress make_spam_user|make_ham_user hooks to collect feedback.
- * @param int $user_id
- * @return void
- */
-function apbct_send_buddypress_user_feedback($user_id)
-{
-    global $apbct;
-
-    // check user rights and if user_id arg
-    if ( !current_user_can('activate_plugins') || empty($user_id) || !is_scalar($user_id)) {
-        return;
-    }
-
-    // get user meta
-    $meta = get_user_meta((int)$user_id);
-
-    // get ct_hash from meta
-    if (empty($meta['ct_hash']) || !isset($meta['ct_hash'][0])) {
-        return;
-    }
-    $ct_hash = $meta['ct_hash'][0];
-    $current_action = current_action();
-
-    // set new solution or return on wrong action
-    if ( $current_action === 'make_spam_user' ) {
-        $feedback_flag = 0;
-    } elseif ( $current_action === 'make_ham_user' ) {
-        $feedback_flag = 1;
-    } else {
-        return;
-    }
-
-    // add message to state, message data will be saved in the ct_feedback() function
-    $apbct->data['bp_feedback_message'] = __('CleanTalk: Your feedback will be sent to the cloud within the next hour. Feel free to contact us via support@cleantalk.org.', 'cleantalk-spam-protect');
-
-    // this will be sent by cron task within an hour
-    ct_feedback($ct_hash, $feedback_flag);
 }
 
 /**
