+++ inc/cleantalk-pluggable.php
@@ -1243,6 +1243,10 @@ function apbct_is_skip_request($ajax = false)
         ) {
             return 'apbct_email_check_exist_post_skip';
         }
+        // BuddyPress has the direct integration
+        if ( apbct_is_plugin_active('buddypress/bp-loader.php') && Post::get('action') === 'messages_send_message' ) {
+            return 'buddypress_messages_send_message';
+        }
     } else {
         /*****************************************/
         /*  Here is non-ajax requests skipping   */
