+++ inc/cleantalk-public-integrations.php
@@ -205,210 +205,6 @@ function ct_woocommerce_wishlist_check($args)
     return $args;
 }
 
-function apbct_integration__buddyPres__getTemplateName(
-    $located,
-    $_template_name,
-    $_template_names,
-    $_template_locations,
-    $_load,
-    $_require_once
-) {
-    global $apbct;
-    preg_match("/\/([a-z-_]+)\/buddypress-functions\.php$/", $located, $matches);
-    $apbct->buddy_press_tmpl = isset($matches[1]) ? $matches[1] : 'unknown';
-}
-
-/**
- * Test BuddyPress activity for spam (post update only)
- *
- * @param bool $is_spam
- * @param object $activity_obj Activity object (\plugins\buddypress\bp-activity\classes\class-bp-activity-activity.php)
- *
- * @return boolean Spam flag
- * @psalm-suppress UnusedVariable
- * @global State $apbct
- */
-function apbct_integration__buddyPres__activityWall($is_spam, $activity_obj = null)
-{
-    global $apbct;
-
-    $allowed_post_actions = array('post_update', 'new_activity_comment');
-
-    if ( ! in_array(Post::get('action'), $allowed_post_actions) ||
-        $activity_obj === null ||
-        ! Post::get('action') ||
-        (isset($activity_obj->privacy) && $activity_obj->privacy == 'media') ||
-        apbct_exclusions_check() ||
-        ! $apbct->settings['forms__contact_forms_test']
-    ) {
-        do_action('apbct_skipped_request', __FILE__ . ' -> ' . __FUNCTION__ . '():' . __LINE__, $_POST);
-
-        return false;
-    }
-
-    $curr_user = get_user_by('id', $activity_obj->user_id);
-
-    //Making a call
-    $base_call_result = apbct_base_call(
-        array(
-            'message'         => is_string($activity_obj->content) ? $activity_obj->content : '',
-            'sender_email'    => $curr_user !== false ? $curr_user->data->user_email : '',
-            'sender_nickname' => $curr_user !== false ? $curr_user->data->user_login : '',
-            'post_info'       => array(
-                'post_url'     => Server::get('HTTP_REFERER'),
-                'comment_type' => 'buddypress_activitywall',
-            ),
-            'sender_info'     => array('sender_url' => null),
-        )
-    );
-
-    if ( isset($base_call_result['ct_result']) ) {
-        $ct_result = $base_call_result['ct_result'];
-
-        if ( $ct_result->allow == 0 ) {
-            add_action('bp_activity_after_save', 'apbct_integration__buddyPres__activityWall_showResponse', 1, 1);
-            $apbct->spam_notification = $ct_result->comment;
-
-            return true;
-        }
-    }
-
-    return $is_spam;
-}
-
-/**
- * Outputs message to AJAX frontend handler
- *
- * @param object $activity_obj Activity object (\plugins\buddypress\bp-activity\classes\class-bp-activity-activity.php)
- *
- * @global State $apbct
- */
-function apbct_integration__buddyPres__activityWall_showResponse($_activity_obj)
-{
-    global $apbct;
-
-    // Legacy template
-    if ( $apbct->buddy_press_tmpl === 'bp-legacy' ) {
-        die('<div id="message" class="error bp-ajax-message"><p>' . $apbct->spam_notification . '</p></div>');
-        // Nouveau template and others
-    } else {
-        @header('Content-Type: application/json; charset=' . get_option('blog_charset'));
-        die(
-            json_encode(
-                array(
-                    'success' => false,
-                    'data'    => array('message' => $apbct->spam_notification),
-                )
-            )
-        );
-    }
-}
-
-/**
- * Public function - Tests new private messages (dialogs)
- *
- * @param object $bp_message_obj
- *
- * @return void with errors if spam has found
- * @psalm-suppress UndefinedClass
- * @psalm-suppress UnusedVariable
- * @global State $apbct
- */
-function apbct_integration__buddyPres__private_msg_check($bp_message_obj)
-{
-    global $apbct;
-
-    //Check for enabled option
-    if (
-        $apbct->settings['comments__bp_private_messages'] == 0 ||
-        apbct_exclusions_check() ||
-        ($apbct->settings['data__protect_logged_in'] == 0 && is_user_logged_in())
-    ) {
-        do_action('apbct_skipped_request', __FILE__ . ' -> ' . __FUNCTION__ . '():' . __LINE__, $_POST);
-
-        return;
-    }
-
-    //Check for quantity of comments
-    $comments_check_number = defined('CLEANTALK_CHECK_COMMENTS_NUMBER')
-        ? CLEANTALK_CHECK_COMMENTS_NUMBER
-        : 3;
-
-    if ( $apbct->settings['comments__check_comments_number'] ) {
-        $args             = array(
-            'user_id'      => $bp_message_obj->sender_id,
-            'box'          => 'sentbox',
-            'type'         => 'all',
-            'limit'        => $comments_check_number,
-            'page'         => null,
-            'search_terms' => '',
-            'meta_query'   => array()
-        );
-        $sentbox_msgs     = BP_Messages_Thread::get_current_threads_for_user($args);
-        $cnt_sentbox_msgs = $sentbox_msgs['total'];
-        $args['box']      = 'inbox';
-        $inbox_msgs       = BP_Messages_Thread::get_current_threads_for_user($args);
-        $cnt_inbox_msgs   = $inbox_msgs['total'];
-
-        if ( ($cnt_inbox_msgs + $cnt_sentbox_msgs) >= $comments_check_number ) {
-            $is_max_comments = true;
-        }
-    }
-
-    $exception_action = false;
-    if ( ! empty($is_max_comments) ) {
-        $exception_action = true;
-    }
-
-    $sender_user_obj = get_user_by('id', $bp_message_obj->sender_id);
-
-    //Making a call
-
-    $event_token = TT::getArrayValueAsString(Post::get('meta'), 'ct_bot_detector_event_token');
-    $base_call_result = apbct_base_call(
-        array(
-            'message'         => $bp_message_obj->subject . " " . $bp_message_obj->message,
-            'sender_email'    => $sender_user_obj !== false ? $sender_user_obj->data->user_email : '',
-            'sender_nickname' => $sender_user_obj !== false ? $sender_user_obj->data->user_login : '',
-            'post_info'       => array(
-                'comment_type' => 'buddypress_comment',
-                'post_url'     => Server::get('HTTP_REFERER'),
-            ),
-            'js_on'           => apbct_js_test(Sanitize::cleanTextField(Cookie::get('ct_checkjs')), true) ?: apbct_js_test(Sanitize::cleanTextField(Post::get('ct_checkjs'))),
-            'sender_info'     => array('sender_url' => null),
-            'exception_action' => $exception_action === true ? 1 : null,
-            'event_token' => $event_token,
-        )
-    );
-
-    if ( isset($base_call_result['ct_result']) ) {
-        $ct_result = $base_call_result['ct_result'];
-
-        if ( $ct_result->allow == 0 ) {
-            if (apbct_is_ajax() ) {
-                wp_send_json_error(
-                    array(
-                        'feedback' => $ct_result->comment,
-                        'type' => 'error'
-                    )
-                );
-            } else {
-                wp_die(
-                    "<h1>"
-                    . __('Spam protection by CleanTalk', 'cleantalk-spam-protect')
-                    . "</h1><h2>" . $ct_result->comment . "</h2>",
-                    '',
-                    array(
-                        'response'       => 403,
-                        "back_link"      => true,
-                        "text_direction" => 'ltr'
-                    )
-                );
-            }
-        }
-    }
-}
-
 
 /**
  * Test default search string for spam
