+++ inc/cleantalk-public.php
@@ -223,14 +223,6 @@ function apbct_init()
         add_action('wpcf7_before_send_mail', 'apbct_form__contactForm7__testSpam', 999);
     }
 
-    // BuddyPress
-    if ( class_exists('BuddyPress') ) {
-        add_action('bp_before_registration_submit_buttons', 'ct_register_form', 1);
-        add_action('messages_message_before_save', 'apbct_integration__buddyPres__private_msg_check', 1);
-        add_filter('bp_signup_validate', 'ct_registration_errors', 1);
-        add_filter('bp_signup_validate', 'ct_check_registration_errors', 999999);
-    }
-
     if ( defined('PROFILEPRESS_SYSTEM_FILE_PATH') ) {
         add_filter('pp_registration_validation', 'ct_registration_errors_ppress', 11, 2);
     }
